// src/clarity-angular/wigfrid/components/FlexGrid/Grid/enum/AutoSizeMode.d.ts
/**
 * Specifies constants that define the row/column auto-sizing behavior.
 */
export declare enum AutoSizeMode {
    /** Autosizing is disabled. */
    None = 0,
    /** Autosizing accounts for header cells. */
    Headers = 1,
    /** Autosizing accounts for data cells. */
    Cells = 2,
    /** Autosizing accounts for header and data cells. */
    Both = 3,
}